requirements.txt
streamlit==1.36.0
pillow==10.3.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.42.0
